builds: AOSP versions
Do not try to keep the Kraken project anymore, rename it to something named AOSP. The channel will still be active for me to post my own modifications. The chat group will be disabled.

I don't want more compression with users, much less with maintainers. I want to work on my fork for my personal use. Even so, the code will be open, after all, nothing is mine, we are open source.

Signed-off-by: Alexandre Rangel <mamutal91@gmail.com>